# account/src/main/resources/openapi/account-ms-openapi.yaml
openapi: 3.0.3
info:
  title: Account API
  version: 1.0.0
  description: API para la gestión de cuentas bancarias asociadas a clientes

servers:
  - url: http://localhost:8081
    description: Servidor local de desarrollo

paths:
  /cuentas:
    post:
      summary: Crear una cuenta bancaria
      description: Crea una cuenta para un cliente validando su existencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Cuenta creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Datos inválidos
        '404':
          description: Cliente no encontrado
    get:
      summary: Listar todas las cuentas bancarias
      responses:
        '200':
          description: Lista de cuentas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'

  /cuentas/{id}:
    get:
      summary: Obtener detalles de una cuenta por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles de la cuenta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Cuenta no encontrada
    delete:
      summary: Eliminar una cuenta bancaria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Cuenta eliminada
        '404':
          description: Cuenta no encontrada

  /cuentas/{cuentaId}/depositar:
    put:
      summary: Realizar un depósito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Depósito realizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '400':
          description: Monto inválido
        '404':
          description: Cuenta no encontrada

  /cuentas/{cuentaId}/retirar:
    put:
      summary: Realizar un retiro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retiro realizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '400':
          description: Monto inválido o retiro no permitido
        '404':
          description: Cuenta no encontrada

components:
  schemas:
    CreateAccountRequest:
      type: object
      properties:
        clientId:
          type: integer
        type:
          type: string
          enum: [SAVINGS, CHECKING]
        initialBalance:
          type: number
          format: double
      required: [clientId, type, initialBalance]

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
        accountNumber:
          type: string
        balance:
          type: number
          format: double
        type:
          type: string
        clientId:
          type: integer

    DepositRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
      required: [amount]

    WithdrawRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
      required: [amount]

    BalanceResponse:
      type: object
      properties:
        id:
          type: integer
        balance:
          type: number
          format: double